poetry_generator/requirements.txt
fastapi
uvicorn
torch
pyyaml
matplotlib
seaborn